services/image_captioning/requirements.txt
transformers==4.18.0
sentencepiece==0.1.94
flask==1.1.1
itsdangerous==2.0.1
gunicorn==19.9.0
requests==2.22.0
sentry-sdk[flask]==0.14.1
healthcheck==1.3.3
jinja2<=3.0.3
Werkzeug<=2.0.3
Pillow==7.0.0
scikit-image==0.16.2
torch==1.7.1